+++ packages/sit-onyx/src/components/OnyxPageLayout/OnyxPageLayout.vue
@@ -15,18 +15,15 @@ const slots = defineSlots<{
   toasts?(): unknown;
 }>();
 
+/**
+ * Determines whether the footer should be below or next to the sidebar.
+ * Does not make a difference when a sidebar is hidden/not filled.
+ */
 const pageModifier = computed(() => {
-  let mode = "";
-  if (!slots.footer && slots.sidebar) mode = "onyx-page--side-main";
-  if (slots.footer && (!slots.sidebar || props.hideSidebar)) mode = "onyx-page--main-footer";
-  if (slots.footer && slots.sidebar) {
-    if (props.footerAsideSidebar) {
-      mode = "onyx-page--side-main-footer-partial";
-    } else {
-      mode = "onyx-page--side-main-footer-full";
-    }
+  if (props.footerAsideSidebar) {
+    return "onyx-page--footer-partial";
   }
-  return mode;
+  return "onyx-page--footer-full";
 });
 </script>
 
@@ -63,24 +60,14 @@ const pageModifier = computed(() => {
     grid-template-columns: auto;
     grid-template-areas: "main";
 
-    &--side-main {
-      grid-template-columns: max-content 1fr;
-      grid-template-areas: "side main";
-    }
-    &--main-footer {
-      grid-template-rows: 1fr max-content;
-      grid-template-areas:
-        "main"
-        "footer";
-    }
-    &--side-main-footer-full {
+    &--footer-full {
       grid-template-columns: max-content 1fr;
       grid-template-rows: 1fr max-content;
       grid-template-areas:
         "side main"
         "footer footer";
     }
-    &--side-main-footer-partial {
+    &--footer-partial {
       grid-template-columns: max-content 1fr;
       grid-template-rows: 1fr max-content;
       grid-template-areas: