+++ apps/alpha-test-app/CHANGELOG.md
@@ -1,5 +1,27 @@
 # alpha-test-app
 
+## 0.1.0-alpha.149
+
+### Patch Changes
+
+- Updated dependencies [8692b19]
+  - sit-onyx@1.0.0-alpha.152
+
+## 0.1.0-alpha.148
+
+### Patch Changes
+
+- Updated dependencies [3c8cf6d]
+  - sit-onyx@1.0.0-alpha.151
+
+## 0.1.0-alpha.147
+
+### Patch Changes
+
+- Updated dependencies [0bdb49a]
+- Updated dependencies [0bdb49a]
+  - sit-onyx@1.0.0-alpha.150
+
 ## 0.1.0-alpha.146
 
 ### Patch Changes