+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @sit-onyx/nuxt
 
+## 1.0.0-alpha.32
+
+### Patch Changes
+
+- Updated dependencies [8692b19]
+  - sit-onyx@1.0.0-alpha.152
+
+## 1.0.0-alpha.31
+
+### Patch Changes
+
+- Updated dependencies [3c8cf6d]
+  - sit-onyx@1.0.0-alpha.151
+
+## 1.0.0-alpha.30
+
+### Patch Changes
+
+- Updated dependencies [0bdb49a]
+- Updated dependencies [0bdb49a]
+  - sit-onyx@1.0.0-alpha.150
+
 ## 1.0.0-alpha.29
 
 ### Patch Changes