+++ packages/storybook-utils/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @sit-onyx/storybook-utils
 
+## 1.0.0-alpha.158
+
+### Patch Changes
+
+- Updated dependencies [8692b19]
+  - sit-onyx@1.0.0-alpha.152
+
+## 1.0.0-alpha.157
+
+### Patch Changes
+
+- Updated dependencies [3c8cf6d]
+  - sit-onyx@1.0.0-alpha.151
+
+## 1.0.0-alpha.156
+
+### Patch Changes
+
+- Updated dependencies [0bdb49a]
+- Updated dependencies [0bdb49a]
+  - sit-onyx@1.0.0-alpha.150
+
 ## 1.0.0-alpha.155
 
 ### Patch Changes